emotion-recognition-system/requirements.txt
# Deep Learning Emotion Recognition System - Dependencies

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
facenet-pytorch>=2.5.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Progress & Utilities
tqdm>=4.65.0
PyYAML>=6.0

# Jupyter Support
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Optional: For better results
# albumentations>=1.3.0  # Advanced augmentations
# timm>=0.9.0  # Additional pretrained models
